+++ lib/src/rust/frb_generated.web.dart
@@ -83,6 +83,9 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
   MultipartPayload dco_decode_box_autoadd_multipart_payload(dynamic raw);
 
   @protected
+  ProxySettings dco_decode_box_autoadd_proxy_settings(dynamic raw);
+
+  @protected
   TlsSettings dco_decode_box_autoadd_tls_settings(dynamic raw);
 
   @protected
@@ -175,6 +178,9 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
   PlatformInt64? dco_decode_opt_box_autoadd_i_64(dynamic raw);
 
   @protected
+  ProxySettings? dco_decode_opt_box_autoadd_proxy_settings(dynamic raw);
+
+  @protected
   TlsSettings? dco_decode_opt_box_autoadd_tls_settings(dynamic raw);
 
   @protected
@@ -182,6 +188,9 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
 
   @protected
   List<(String, String)>? dco_decode_opt_list_record_string_string(dynamic raw);
+
+  @protected
+  ProxySettings dco_decode_proxy_settings(dynamic raw);
 
   @protected
   (HttpHeaderName, String) dco_decode_record_http_header_name_string(
@@ -267,6 +276,10 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
 
   @protected
   MultipartPayload sse_decode_box_autoadd_multipart_payload(
+      SseDeserializer deserializer);
+
+  @protected
+  ProxySettings sse_decode_box_autoadd_proxy_settings(
       SseDeserializer deserializer);
 
   @protected
@@ -368,6 +381,10 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
   PlatformInt64? sse_decode_opt_box_autoadd_i_64(SseDeserializer deserializer);
 
   @protected
+  ProxySettings? sse_decode_opt_box_autoadd_proxy_settings(
+      SseDeserializer deserializer);
+
+  @protected
   TlsSettings? sse_decode_opt_box_autoadd_tls_settings(
       SseDeserializer deserializer);
 
@@ -378,6 +395,9 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
   @protected
   List<(String, String)>? sse_decode_opt_list_record_string_string(
       SseDeserializer deserializer);
+
+  @protected
+  ProxySettings sse_decode_proxy_settings(SseDeserializer deserializer);
 
   @protected
   (HttpHeaderName, String) sse_decode_record_http_header_name_string(
@@ -479,6 +499,10 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
   @protected
   void sse_encode_box_autoadd_multipart_payload(
       MultipartPayload self, SseSerializer serializer);
+
+  @protected
+  void sse_encode_box_autoadd_proxy_settings(
+      ProxySettings self, SseSerializer serializer);
 
   @protected
   void sse_encode_box_autoadd_tls_settings(
@@ -592,6 +616,10 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
       PlatformInt64? self, SseSerializer serializer);
 
   @protected
+  void sse_encode_opt_box_autoadd_proxy_settings(
+      ProxySettings? self, SseSerializer serializer);
+
+  @protected
   void sse_encode_opt_box_autoadd_tls_settings(
       TlsSettings? self, SseSerializer serializer);
 
@@ -602,6 +630,9 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
   @protected
   void sse_encode_opt_list_record_string_string(
       List<(String, String)>? self, SseSerializer serializer);
+
+  @protected
+  void sse_encode_proxy_settings(ProxySettings self, SseSerializer serializer);
 
   @protected
   void sse_encode_record_http_header_name_string(