+++ rust/src/frb_generated.rs
@@ -499,6 +499,8 @@ impl SseDecode for crate::api::client::ClientSettings {
         let mut var_timeout = <Option<chrono::Duration>>::sse_decode(deserializer);
         let mut var_connectTimeout = <Option<chrono::Duration>>::sse_decode(deserializer);
         let mut var_throwOnStatusCode = <bool>::sse_decode(deserializer);
+        let mut var_proxySettings =
+            <Option<crate::api::client::ProxySettings>>::sse_decode(deserializer);
         let mut var_tlsSettings =
             <Option<crate::api::client::TlsSettings>>::sse_decode(deserializer);
         return crate::api::client::ClientSettings {
@@ -506,6 +508,7 @@ impl SseDecode for crate::api::client::ClientSettings {
             timeout: var_timeout,
             connect_timeout: var_connectTimeout,
             throw_on_status_code: var_throwOnStatusCode,
+            proxy_settings: var_proxySettings,
             tls_settings: var_tlsSettings,
         };
     }
@@ -964,6 +967,19 @@ impl SseDecode for Option<i64> {
     }
 }
 
+impl SseDecode for Option<crate::api::client::ProxySettings> {
+    // Codec=Sse (Serialization based), see doc to use other codecs
+    fn sse_decode(deserializer: &mut flutter_rust_bridge::for_generated::SseDeserializer) -> Self {
+        if (<bool>::sse_decode(deserializer)) {
+            return Some(<crate::api::client::ProxySettings>::sse_decode(
+                deserializer,
+            ));
+        } else {
+            return None;
+        }
+    }
+}
+
 impl SseDecode for Option<crate::api::client::TlsSettings> {
     // Codec=Sse (Serialization based), see doc to use other codecs
     fn sse_decode(deserializer: &mut flutter_rust_bridge::for_generated::SseDeserializer) -> Self {
@@ -994,6 +1010,17 @@ impl SseDecode for Option<Vec<(String, String)>> {
         } else {
             return None;
         }
+    }
+}
+
+impl SseDecode for crate::api::client::ProxySettings {
+    // Codec=Sse (Serialization based), see doc to use other codecs
+    fn sse_decode(deserializer: &mut flutter_rust_bridge::for_generated::SseDeserializer) -> Self {
+        let mut inner = <i32>::sse_decode(deserializer);
+        return match inner {
+            0 => crate::api::client::ProxySettings::NoProxy,
+            _ => unreachable!("Invalid variant for ProxySettings: {}", inner),
+        };
     }
 }
 
@@ -1195,6 +1222,7 @@ impl flutter_rust_bridge::IntoDart for crate::api::client::ClientSettings {
             self.timeout.into_into_dart().into_dart(),
             self.connect_timeout.into_into_dart().into_dart(),
             self.throw_on_status_code.into_into_dart().into_dart(),
+            self.proxy_settings.into_into_dart().into_dart(),
             self.tls_settings.into_into_dart().into_dart(),
         ]
         .into_dart()
@@ -1579,6 +1607,26 @@ impl flutter_rust_bridge::IntoIntoDart<crate::api::http::MultipartValue>
     }
 }
 // Codec=Dco (DartCObject based), see doc to use other codecs
+impl flutter_rust_bridge::IntoDart for crate::api::client::ProxySettings {
+    fn into_dart(self) -> flutter_rust_bridge::for_generated::DartAbi {
+        match self {
+            Self::NoProxy => 0.into_dart(),
+            _ => unreachable!(),
+        }
+    }
+}
+impl flutter_rust_bridge::for_generated::IntoDartExceptPrimitive
+    for crate::api::client::ProxySettings
+{
+}
+impl flutter_rust_bridge::IntoIntoDart<crate::api::client::ProxySettings>
+    for crate::api::client::ProxySettings
+{
+    fn into_into_dart(self) -> crate::api::client::ProxySettings {
+        self
+    }
+}
+// Codec=Dco (DartCObject based), see doc to use other codecs
 impl flutter_rust_bridge::IntoDart for crate::api::error::RhttpError {
     fn into_dart(self) -> flutter_rust_bridge::for_generated::DartAbi {
         match self {
@@ -1737,6 +1785,7 @@ impl SseEncode for crate::api::client::ClientSettings {
         <Option<chrono::Duration>>::sse_encode(self.timeout, serializer);
         <Option<chrono::Duration>>::sse_encode(self.connect_timeout, serializer);
         <bool>::sse_encode(self.throw_on_status_code, serializer);
+        <Option<crate::api::client::ProxySettings>>::sse_encode(self.proxy_settings, serializer);
         <Option<crate::api::client::TlsSettings>>::sse_encode(self.tls_settings, serializer);
     }
 }
@@ -2169,6 +2218,16 @@ impl SseEncode for Option<i64> {
     }
 }
 
+impl SseEncode for Option<crate::api::client::ProxySettings> {
+    // Codec=Sse (Serialization based), see doc to use other codecs
+    fn sse_encode(self, serializer: &mut flutter_rust_bridge::for_generated::SseSerializer) {
+        <bool>::sse_encode(self.is_some(), serializer);
+        if let Some(value) = self {
+            <crate::api::client::ProxySettings>::sse_encode(value, serializer);
+        }
+    }
+}
+
 impl SseEncode for Option<crate::api::client::TlsSettings> {
     // Codec=Sse (Serialization based), see doc to use other codecs
     fn sse_encode(self, serializer: &mut flutter_rust_bridge::for_generated::SseSerializer) {
@@ -2196,6 +2255,21 @@ impl SseEncode for Option<Vec<(String, String)>> {
         if let Some(value) = self {
             <Vec<(String, String)>>::sse_encode(value, serializer);
         }
+    }
+}
+
+impl SseEncode for crate::api::client::ProxySettings {
+    // Codec=Sse (Serialization based), see doc to use other codecs
+    fn sse_encode(self, serializer: &mut flutter_rust_bridge::for_generated::SseSerializer) {
+        <i32>::sse_encode(
+            match self {
+                crate::api::client::ProxySettings::NoProxy => 0,
+                _ => {
+                    unimplemented!("");
+                }
+            },
+            serializer,
+        );
     }
 }