+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
 ## 0.4.0
 
-- feat: add `RHttpCompatibleClient` that is compatible with the [http](https://pub.dev/packages/http) package
+- feat: add `RhttpCompatibleClient` that is compatible with the [http](https://pub.dev/packages/http) package
 
 ## 0.3.2